Pressure and Temperature
Pressure and Temperature sensors are both transmitting their states to
the serial and they aren't affecting each other, but are also changing
when we change it. So they are working separately on the same circuit
board and on the same code. We combined the pressure sensor code and the
temperature sensor code.